Crate ccthw_ash_allocator

A general purpose Vulkan Memory allocator, written from scratch the hard way.

Structs

Allocation

A GPU memory allocation.

AllocationRequirements

All supported memory requirements.

DeviceAllocator

A GPU memory allocator which always allocates memory directly from the device.

MemoryAllocator

The top-level interface for allocating GPU memory.

MemoryProperties

PageSuballocator

TraceAllocator

An allocator decorator which tracks metrics and generates a report for all allocations made to the wrapped allocator.

Enums

AllocatorError

DedicatedResourceHandle

A copy of the resource handle associated with an allocation.

Traits

ComposableAllocator