Crate ccthw_ash_allocator

A general purpose Vulkan Memory allocator, written from scratch the hard way.

Structs

Allocation

A GPU memory allocation.

AllocationRequirements

All supported memory requirements.

DedicatedAllocator

An allocator which correctly handles allocations which prefer or require dedicated allocations.

DeviceAllocator

A GPU memory allocator which always allocates memory directly from the device.

FakeAllocator

A fake implementation of a composable memory allocator which keeps track of all requested memory allocations.

MemoryAllocator

The top-level interface for allocating GPU memory.

MemoryProperties

MemoryTypePoolAllocator

PageSuballocator

PoolAllocator

SizedAllocator

An allocator which composes over two other allocators. When a request is below the trigger size, it is sent to the first alloctor, otherwise it is sent to the second allocator.

TraceAllocator

An allocator decorator which tracks metrics and generates a report for all allocations made to the wrapped allocator.

Enums

AllocatorError

DedicatedResourceHandle

A copy of the resource handle associated with an allocation.

Traits

ComposableAllocator

Functions

create_system_allocator⚠

Create an opinionated system allocator for GPU memoy.

into_shared

Move an composable allocator into a Rc RefCell.