Trait ccthw_ash_allocator::ComposableAllocator

pub trait ComposableAllocator {
    // Required methods
    unsafe fn allocate(
        &mut self,
        allocation_requirements: AllocationRequirements
    ) -> Result<Allocation, AllocatorError>;
    unsafe fn free(&mut self, allocation: Allocation);
}

Required Methods

unsafe fn allocate( &mut self, allocation_requirements: AllocationRequirements ) -> Result<Allocation, AllocatorError>

Allocate GPU memory based on the given requirements.

Safety

Unsafe because memory must be freed before the device is destroyed.

unsafe fn free(&mut self, allocation: Allocation)

Return a GPU memory allocation to the device.

Safety

Unsafe because:

memory must be freed by the application before the device is destroyed

the application is responsible for synchronizing access to device memory. It is an error to free memory while ongoing GPU operations are still referencing it.